requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.22.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic>=0.75.0
anyio==4.12.0
appdirs==1.4.4
asgiref==3.11.0
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.4.0
backoff==2.2.1
banks==2.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
build==1.3.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb>=0.5.17
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography>=45.0.0
dataclasses-json==0.6.7
datasets>=2.14.0
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
docker-pycreds==0.4.0
durationpy==0.10
einops==0.7.0
fastapi==0.127.0
python-multipart>=0.0.18
filelock==3.20.1
filetype==1.2.0
Flask==3.1.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.4
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
google-generativeai==0.5.4
googleapis-common-protos==1.72.0
greenlet==3.3.0
griffe==1.15.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.4.0
importlib_resources==6.5.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
# LlamaIndex Core - Upgraded for GPT-5 support (Phase 1)
# Core and LLM providers - pinned for GPT-5 support
llama-index-core>=0.14.5,<0.15
llama-index-llms-openai>=0.6.0
llama-index-llms-anthropic>=0.10.0
llama-index-llms-gemini>=0.6.0
llama-index-llms-ollama>=0.9.0
llama-index-llms-groq>=0.3.0
groq>=0.11.0  # SDK for RateLimitError exception handling
# Embeddings
llama-index-embeddings-huggingface>=0.6.0
llama-index-embeddings-openai>=0.5.0
# Vector stores and retrieval
llama-index-vector-stores-postgres>=0.7.0
llama-index-retrievers-bm25>=0.6.0
llama-index-vector-stores-chroma>=0.4.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.109.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pgvector>=0.3.6,<1.0.0
pillow==10.4.0
platformdirs==4.5.1
pluggy==1.6.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==4.25.8
psutil==7.1.3
psycopg2-binary==2.9.11
pulsar-client==3.8.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.26.7
pymupdf4llm>=0.0.17
PyMySQL>=1.1.0
pyparsing==3.3.1
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==2.7.0
sentry-sdk==2.48.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
starlette==0.50.0
striprtf==0.0.26
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typer==0.20.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.40.0
uvloop==0.22.1
wandb==0.16.6
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
tavily-python>=0.5.0

# Tree-sitter (CodeLoom AST parsing)
tree-sitter>=0.23.0
tree-sitter-python>=0.23.0
tree-sitter-javascript>=0.23.0
tree-sitter-typescript>=0.23.0
tree-sitter-java>=0.23.0
tree-sitter-c-sharp>=0.23.0
